+++ tests/test_stiffness_checker.py
@@ -60,12 +60,14 @@ def repetitive_test_stiffness_checker(frame_file_path, parsed_pt_loads=None, inc
     # without reinit
     st_time = time.time()
     for _ in range(n_attempts):
-        sol_success = sc.solve(existing_ids)
+        shuffled_existing_ids = existing_ids.copy()
+        random.shuffle(shuffled_existing_ids)
+        sol_success = sc.solve(shuffled_existing_ids)
         tmp_success, tmp_nD, tmp_fR, tmp_eR = sc.get_solved_results()
 
         assert tmp_success == success and sol_success == success, 'max_trans {}, max_rot {}, max_trans_id {}, max_rot_id {}'.format( \
             *sc.get_max_nodal_deformation())
-        for fnid in sc.get_element_connected_node_ids(existing_ids=existing_ids, fix_node_only=True):
+        for fnid in sc.get_element_connected_node_ids(existing_ids=shuffled_existing_ids, fix_node_only=True):
             assert_equal(tmp_nD[fnid], np.zeros(6))
 
         for i in nD.keys() : assert_almost_equal(nD[i], tmp_nD[i]) 
@@ -83,11 +85,13 @@ def repetitive_test_stiffness_checker(frame_file_path, parsed_pt_loads=None, inc
         sc_new.set_loads(point_loads=parsed_pt_loads, include_self_weight=include_sw)
         sc_new.set_nodal_displacement_tol(trans_tol=trans_tol)
 
-        sol_success = sc_new.solve(existing_ids)
+        shuffled_existing_ids = existing_ids.copy()
+        random.shuffle(shuffled_existing_ids)
+        sol_success = sc_new.solve(shuffled_existing_ids)
         tmp_success, tmp_nD, tmp_fR, tmp_eR = sc_new.get_solved_results()
         assert tmp_success == success and sol_success == success, 'max_trans {}, max_rot {}, max_trans_id {}, max_rot_id {}'.format( \
             *sc_new.get_max_nodal_deformation())
-        for fnid in sc_new.get_element_connected_node_ids(existing_ids=existing_ids, fix_node_only=True):
+        for fnid in sc_new.get_element_connected_node_ids(existing_ids=shuffled_existing_ids, fix_node_only=True):
             assert_equal(tmp_nD[fnid], np.zeros(6))
 
         for i in nD.keys() : assert_almost_equal(nD[i], tmp_nD[i]) 
@@ -108,7 +112,7 @@ def test_stiffness_checker_solve_consistency(test_case, load_case):
         file_name = 'tower_3D.json'
         load_file_name = 'tower_3D_load_case.json'
         
-        # some parts are floating in this partial ids
+        # ! some parts are floating in this partial ids
         failed_existing_ids = [0, 4, 7, 8, 9] 
         if load_case == 'self_weight' or load_case == 'self_weight+point_load':
             success_existing_ids = [1, 12, 5, 10, 4, 11, 13, 3, 2, 0, 7]
@@ -147,7 +151,7 @@ def test_stiffness_checker_solve_consistency(test_case, load_case):
     
 
     # repetitive tests
-    n_attempts = 10
+    n_attempts = 100
     
     print('################\nfull solve success checks')
     repetitive_test_stiffness_checker(json_path, parsed_pt_loads=parsed_pt_loads, include_sw=include_sw, \
@@ -452,12 +456,21 @@ def test_nodal_equilibrium(test_case, load_case):
     repetitive_test_equilibrium(json_path, parsed_pt_loads=parsed_pt_loads, include_sw=include_sw, \
         existing_ids=failed_existing_ids, expect_partial_ids_success=False)
 
-
 def repetitive_check_gravity_validity(frame_file_path, existing_ids=[], expect_partial_ids_success=True):
     sc = StiffnessChecker(json_file_path=frame_file_path)
-    sc.set_loads(include_self_weight=True)
 
-    sol_success = sc.solve(existing_ids, eid_sanity_check=True)
+    print('\n################\n [0,0,-100] gravity direction')
+    sc.set_loads(include_self_weight=True, gravity_direction=[0,0,-100])
+    sc.solve(existing_ids, eid_sanity_check=True)
+    compliance_big = sc.get_compliance()
+
+    print('\n################\n [0,0,-1] gravity direction')
+    sc.set_loads(include_self_weight=True, gravity_direction=[0,0,-1])
+    sc.solve(existing_ids, eid_sanity_check=True)
+    compliance_small = sc.get_compliance()
+
+    print('big comp: {}, small comp: {}'.format(compliance_big, compliance_small))
+    assert_almost_equal(compliance_big, compliance_small * 1e4)
 
     success, nD, fR, eR = sc.get_solved_results()
 
@@ -579,8 +592,9 @@ def test_uniformly_distributed_load_with_gravity(test_case, existing_e_ids):
         assert_almost_equal(ul_eR[e_id][0], sw_eR[e_id][0])
         assert_almost_equal(ul_eR[e_id][1], sw_eR[e_id][1])
 
-@pytest.mark.ignore(reason='not fully developed')
+# @pytest.mark.ignore(reason='not fully developed')
 @pytest.mark.uniform_load_check
+@pytest.mark.analy_compare
 def test_uniformly_distributed_load_with_analytical_solution():
     """ analytical example in 
             Matrix Structural Analysis 2rd edition, McGuire, Gallagher, Ziemian
@@ -598,27 +612,42 @@ def test_uniformly_distributed_load_with_analytical_solution():
     load_json_path = os.path.join(root_dir, 'test_data', load_file_name)
 
     sc = StiffnessChecker(json_file_path=json_path)
-    
     point_load, uniform_element_load, include_sw = read_load_case_json(load_json_path)
     sc.set_loads(point_loads=point_load, uniform_distributed_load=uniform_element_load, include_self_weight=include_sw)
+    sc.set_nodal_displacement_tol(trans_tol=0.024, rot_tol=0.006)
 
-    sc.solve()
-    pass_criteria, nD, fR, eR = sc.get_solved_results()
+    def compare_analytical_sol(pass_criteria, nD, fR, eR, nodal_loads, check_decimal=1):
+        assert_equal(nodal_loads[1][2], -12.5)
+        assert_equal(nodal_loads[1][3], 0.0)
+        assert_equal(nodal_loads[1][4], -6.250)
+        print('{} \t {}'.format(fR[2][0:3], fR[2][3:6]))
 
-    nodal_loads = sc.get_nodal_loads()
-    assert_equal(nodal_loads[1][2], -12.5)
-    assert_equal(nodal_loads[1][3], 0.0)
-    assert_equal(nodal_loads[1][4], -6.250)
-    
-    check_decimal = 2
-    assert not pass_criteria
-    assert_equal(nD[0], [0] * 6)
-    assert_almost_equal(nD[1], [0, 0, -0.02237, 4.195*1e-3, 5.931*1e-3, 0], decimal=check_decimal)
-    assert_equal(nD[2], [0] * 6)
+        assert pass_criteria
+        assert_equal(nD[0], [0] * 6)
+        assert_almost_equal(nD[1], [0, 0, -0.02237, 4.195*1e-3, 5.931*1e-3, 0], decimal=check_decimal)
+        assert_equal(nD[2], [0] * 6)
+        assert_almost_equal(fR[0], [0, 0, 14.74, -6.45*1e-3, -36.21, 0], decimal=check_decimal)
+        # assert_almost_equal(fR[2], [0, 0, 5.25, -41.94, -17.11*1e-3, 0], decimal=check_decimal)
 
-    check_decimal = 1
-    assert_almost_equal(fR[0], [0, 0, 14.74, -6.45*1e-3, -36.21, 0], decimal=check_decimal)
-    assert_almost_equal(fR[2], [0, 0, 5.25, -41.94, -17.11*1e-3, 0], decimal=check_decimal)
+    test_iters = 100
+    print('compare analytical res: w/o reinit')
+    for _ in range(test_iters):
+        sc.solve()
+        pass_criteria, nD, fR, eR = sc.get_solved_results()
+        nodal_loads = sc.get_nodal_loads()    
+        compare_analytical_sol(pass_criteria, nD, fR, eR, nodal_loads)
+
+    print('compare analytical res: w reinit')
+    for _ in range(test_iters):
+        re_sc = StiffnessChecker(json_file_path=json_path)
+        point_load, uniform_element_load, include_sw = read_load_case_json(load_json_path)
+        re_sc.set_loads(point_loads=point_load, uniform_distributed_load=uniform_element_load, include_self_weight=include_sw)
+        re_sc.set_nodal_displacement_tol(trans_tol=0.024, rot_tol=0.006)
+
+        re_sc.solve()
+        pass_criteria, nD, fR, eR = re_sc.get_solved_results()
+        nodal_loads = re_sc.get_nodal_loads()    
+        compare_analytical_sol(pass_criteria, nD, fR, eR, nodal_loads)
 
 
 # @pytest.mark.ii