+++ CHANGELOG.rst
@@ -10,6 +10,14 @@ Changelog
 .. ^, for subsubsections
 .. ", for paragraphs
 
+0.3.0
+----------
+
+Changed
+^^^^^^^
+
+- Changed `try/catch` in the C++ file parsing to `std::throw` 
+
 0.2.0
 -----
 
@@ -31,6 +39,7 @@ Added
 - add grasshopper examples - parse/save files, karamba comparsion, solve/get result in GH via ghpython-remote
 - supports material / cross sectional properties for each element. 
 - supports uniformly distributed load
+- add gravity magnitude and direction
 
 0.1.0
 -----